# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "description": "",
        "operationId": "signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "summary": "Create a new user account",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "",
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponseDto"
                }
              }
            }
          }
        },
        "summary": "Login with email and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Access token refreshed successfully",
        "operationId": "refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponseDto"
                }
              }
            }
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Logged out successfully",
        "operationId": "logout",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logged out successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Logout from the current session",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/features": {
      "post": {
        "description": "Creates a new feature in the system. Admin access required.",
        "operationId": "createFeature",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeatureDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a new feature in the system. Admin access required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Create a new feature",
        "tags": [
          "Features"
        ]
      },
      "get": {
        "description": "Retrieves all features in the system. Admin access required.",
        "operationId": "getAllFeatures",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Retrieves all features in the system. Admin access required.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get all features",
        "tags": [
          "Features"
        ]
      }
    },
    "/features/{name}": {
      "get": {
        "description": "Retrieves a feature by its name. Admin access required.",
        "operationId": "getFeatureByName",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieves a feature by its name. Admin access required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get a feature by name",
        "tags": [
          "Features"
        ]
      },
      "put": {
        "description": "Updates an existing feature by name. Admin access required.",
        "operationId": "updateFeature",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeatureDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updates an existing feature by name. Admin access required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Update a feature",
        "tags": [
          "Features"
        ]
      },
      "delete": {
        "description": "Deletes a feature by name. Admin access required.",
        "operationId": "deleteFeature",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deletes a feature by name. Admin access required."
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Delete a feature",
        "tags": [
          "Features"
        ]
      }
    },
    "/features/website": {
      "post": {
        "description": "Enables and configures a feature for a specific website. Admin or website owner access required.",
        "operationId": "enableFeatureForWebsite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebsiteFeatureDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Enables and configures a feature for a specific website. Admin or website owner access required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebsiteFeature"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Enable/Configure a feature for a website",
        "tags": [
          "Features"
        ]
      }
    },
    "/features/website/{websiteId}": {
      "get": {
        "description": "Retrieves all features configured for a specific website. Admin or website owner access required.",
        "operationId": "getWebsiteFeatures",
        "parameters": [
          {
            "name": "websiteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieves all features configured for a specific website. Admin or website owner access required.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WebsiteFeature"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get features for a website",
        "tags": [
          "Features"
        ]
      }
    },
    "/features/access-level": {
      "post": {
        "description": "Sets permissions for a feature based on access level. Admin or website owner access required.",
        "operationId": "setFeatureAccessLevel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessLevelFeatureDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sets permissions for a feature based on access level. Admin or website owner access required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessLevelFeature"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Set access level permissions for a feature",
        "tags": [
          "Features"
        ]
      }
    },
    "/features/access-level/{featureName}": {
      "get": {
        "description": "Retrieves all access level permissions for a specific feature. Admin or website owner access required.",
        "operationId": "getFeatureAccessLevels",
        "parameters": [
          {
            "name": "featureName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieves all access level permissions for a specific feature. Admin or website owner access required.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccessLevelFeature"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get access levels for a feature",
        "tags": [
          "Features"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pulse Analytics",
    "description": "Pulse Analytics API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SignInResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "user",
          "accessToken",
          "refreshToken"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "RefreshTokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message"
          }
        },
        "required": [
          "message"
        ]
      },
      "CreateFeatureDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Unique name identifier for the feature",
            "example": "realtime-analytics"
          },
          "description": {
            "type": "string",
            "description": "Description of what the feature does",
            "example": "Provides real-time analytics and visitor tracking"
          },
          "category": {
            "type": "string",
            "description": "Category the feature belongs to",
            "example": "analytics",
            "enum": [
              "analytics",
              "reporting",
              "settings",
              "user-management"
            ]
          },
          "isGloballyAvailable": {
            "type": "boolean",
            "description": "Whether the feature is available to all websites by default",
            "example": true,
            "default": true
          }
        },
        "required": [
          "name",
          "description",
          "category"
        ]
      },
      "Feature": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the feature",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "description": "Unique name identifier for the feature",
            "example": "realtime-analytics"
          },
          "description": {
            "type": "string",
            "description": "Description of what the feature does",
            "example": "Provides real-time analytics and visitor tracking"
          },
          "category": {
            "type": "string",
            "description": "Category the feature belongs to",
            "example": "analytics",
            "enum": [
              "analytics",
              "reporting",
              "settings",
              "user-management"
            ]
          },
          "isGloballyAvailable": {
            "type": "boolean",
            "description": "Whether the feature is available to all websites by default",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the feature was created",
            "example": "2025-03-01T00:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the feature was last updated",
            "example": "2025-03-01T00:00:00Z"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "category",
          "isGloballyAvailable",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateFeatureDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Updated description of what the feature does",
            "example": "Provides enhanced real-time analytics and visitor tracking"
          },
          "category": {
            "type": "string",
            "description": "Updated category the feature belongs to",
            "example": "analytics",
            "enum": [
              "analytics",
              "reporting",
              "settings",
              "user-management"
            ]
          },
          "isGloballyAvailable": {
            "type": "boolean",
            "description": "Whether the feature is available to all websites by default",
            "example": true
          }
        }
      },
      "WebsiteFeatureDto": {
        "type": "object",
        "properties": {
          "websiteId": {
            "type": "string",
            "description": "UUID of the website to enable the feature for",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "featureName": {
            "type": "string",
            "description": "Name of the feature to enable",
            "example": "realtime-analytics"
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Whether the feature is enabled for this website",
            "example": true
          },
          "configuration": {
            "type": "object",
            "description": "Website-specific configuration for the feature",
            "example": {
              "refreshInterval": 30,
              "maxDataPoints": 100,
              "enableNotifications": true
            }
          }
        },
        "required": [
          "websiteId",
          "featureName",
          "isEnabled"
        ]
      },
      "Website": {
        "type": "object",
        "properties": {}
      },
      "WebsiteFeature": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the website feature relationship",
            "example": "660e8400-e29b-41d4-a716-446655440000"
          },
          "website": {
            "description": "The website this feature is enabled for",
            "allOf": [
              {
                "$ref": "#/components/schemas/Website"
              }
            ]
          },
          "feature": {
            "description": "The feature that is enabled for the website",
            "allOf": [
              {
                "$ref": "#/components/schemas/Feature"
              }
            ]
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Whether the feature is enabled for this website",
            "example": true
          },
          "configuration": {
            "type": "object",
            "description": "Website-specific feature configuration",
            "example": {
              "refreshInterval": 30,
              "maxDataPoints": 100,
              "enableNotifications": true
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the feature was enabled for this website",
            "example": "2025-03-01T00:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the website feature configuration was last updated",
            "example": "2025-03-01T00:00:00Z"
          }
        },
        "required": [
          "id",
          "website",
          "feature",
          "isEnabled",
          "configuration",
          "createdAt",
          "updatedAt"
        ]
      },
      "AccessLevelFeatureDto": {
        "type": "object",
        "properties": {
          "featureName": {
            "type": "string",
            "description": "Name of the feature to set access level for",
            "example": "realtime-analytics"
          },
          "accessLevel": {
            "type": "string",
            "description": "Access level to set permissions for",
            "example": "ADMIN",
            "enum": [
              "owner",
              "admin",
              "analyst",
              "viewer"
            ]
          },
          "permissions": {
            "type": "object",
            "description": "Permission settings for this feature and access level",
            "example": {
              "canView": true,
              "canEdit": true,
              "canManage": false
            }
          }
        },
        "required": [
          "featureName",
          "accessLevel",
          "permissions"
        ]
      },
      "AccessLevelFeature": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the access level feature relationship",
            "example": "880e8400-e29b-41d4-a716-446655440000"
          },
          "accessLevel": {
            "type": "string",
            "description": "Access level for this feature permission set",
            "example": "ADMIN",
            "enum": [
              "owner",
              "admin",
              "analyst",
              "viewer"
            ]
          },
          "feature": {
            "description": "The feature these permissions apply to",
            "allOf": [
              {
                "$ref": "#/components/schemas/Feature"
              }
            ]
          },
          "permissions": {
            "type": "object",
            "description": "Permission settings for this feature and access level",
            "example": {
              "canView": true,
              "canEdit": true,
              "canManage": false
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date when this permission set was created",
            "example": "2025-03-01T00:00:00Z"
          }
        },
        "required": [
          "id",
          "accessLevel",
          "feature",
          "permissions",
          "createdAt"
        ]
      },
...
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}